Added Weapon & altered Ragdoll
I'm setting up a mechanic of swinging a sword and that if the sword hits a certain part of the AI, it will launch it from that limb and in the direction of the swinging sword.

Only have the sword locked in place rn and FP character added to scene. Was stuck for a while on the weapon and character colliding which was ruining the fps functionality.